# openapi.yaml
openapi: 3.0.0
info:
  title: API de Equipos de Fútbol
  version: 1.0.0
  description: API para gestionar equipos de fútbol
servers:
  - url: http://localhost:3000
paths:
  /equipos:
    get:
      summary: Obtener todos los equipos
      responses:
        '200':
          description: Lista de equipos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipo'
    post:
      summary: Crear un equipo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipo'
      responses:
        '201':
          description: Equipo creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipo'
  /equipos/{id}:
    get:
      summary: Obtener un equipo por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Equipo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipo'
        '404':
          description: Equipo no encontrado
    put:
      summary: Actualizar un equipo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipo'
      responses:
        '200':
          description: Equipo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipo'
        '404':
          description: Equipo no encontrado
    delete:
      summary: Eliminar un equipo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Equipo eliminado
        '404':
          description: Equipo no encontrado
components:
  schemas:
    Equipo:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        ciudad:
          type: string
        añoFundacion:
          type: integer
        esActivo:
          type: boolean
        categoria:
          type: string
      required:
        - nombre
        - categoria